Export missing classes and aliases
